Extracts out get_named_capture out of utils
